# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a table top game.",
        "title": "Table Top API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/weapons": {
            "get": {
                "description": "List all weapons",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weapons"
                ],
                "summary": "List all weapons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Weapon"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new weapon along with its combos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weapons"
                ],
                "summary": "Create a new weapon",
                "parameters": [
                    {
                        "description": "Weapon",
                        "name": "weapon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/weapons.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Weapon"
                        }
                    }
                }
            }
        },
        "/weapons/{id}": {
            "get": {
                "description": "Get a weapon by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weapons"
                ],
                "summary": "Get a weapon by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Weapon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Weapon"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a weapon",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weapons"
                ],
                "summary": "Update a weapon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Weapon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Weapon",
                        "name": "weapon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/weapons.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Weapon"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a weapon by ID",
                "tags": [
                    "Weapons"
                ],
                "summary": "Delete a weapon by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Weapon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Weapon": {
            "type": "object",
            "properties": {
                "character_id": {
                    "type": "string"
                },
                "count_cards": {
                    "type": "integer"
                },
                "defense": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "weapon_combo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WeaponCombo"
                    }
                }
            }
        },
        "models.WeaponCombo": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.WeaponComboType"
                },
                "weapon_id": {
                    "type": "string"
                }
            }
        },
        "models.WeaponComboType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "FlexSlot",
                "CombatSkillSlot",
                "AbilitySlot"
            ]
        },
...
        "weapons.CreateRequest": {
            "type": "object",
            "properties": {
                "character_id": {
                    "type": "string"
                },
                "count_cards": {
                    "type": "integer"
                },
                "defense": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "weapon_combo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/weapons.WeaponCombo"
                    }
                }
            }
        },
        "weapons.UpdateRequest": {
            "type": "object",
            "properties": {
                "character_id": {
                    "type": "string"
                },
                "count_cards": {
                    "type": "integer"
                },
                "defense": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "weapons.WeaponCombo": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.WeaponComboType"
                }
            }
        }
    }
}